Allow request timeout values to be of type long

+++ src/main/java/mousio/etcd4j/requests/EtcdKeyDeleteRequest.java
@@ -65,7 +65,7 @@ public class EtcdKeyDeleteRequest extends EtcdKeyRequest {
     return this;
   }
 
-  @Override public EtcdKeyDeleteRequest timeout(int timeout, TimeUnit unit) {
+  @Override public EtcdKeyDeleteRequest timeout(long timeout, TimeUnit unit) {
     super.timeout(timeout, unit);
     return this;
   }
